Python/3_python_api_dev/1.backups_and_course_source/backup_3_at_017/requirements.txt
fastapi
# sqlalchemy
psycopg2-binary
